MB-12120 Move shipment orchestrator to be in an orchestrators dir
Also rename package to make it easier to read.
