+++ pkg/services/orchestrators/shipment/rules_test.go
@@ -1,4 +1,4 @@
-package shipmentorchestrator
+package shipment
 
 import (
 	"fmt"
