+++ pkg/services/orchestrators/shipment/shipment_creator_test.go
@@ -1,4 +1,4 @@
-package shipmentorchestrator
+package shipment
 
 import (
 	"fmt"
