+++ pkg/services/orchestrators/shipment/shipment_updater.go
@@ -1,4 +1,4 @@
-package shipmentorchestrator
+package shipment
 
 import (
 	"github.com/transcom/mymove/pkg/appcontext"
