+++ pkg/services/orchestrators/shipment/shipment_updater_test.go
@@ -1,4 +1,4 @@
-package shipmentorchestrator
+package shipment
 
 import (
 	"fmt"
