+++ pkg/services/orchestrators/shipment/validation_test.go
@@ -1,4 +1,4 @@
-package shipmentorchestrator
+package shipment
 
 import (
 	"github.com/transcom/mymove/pkg/apperror"
